+++ com.unity.render-pipelines.core/Runtime/Volume/VolumeStack.cs
@@ -66,7 +66,10 @@ namespace UnityEngine.Rendering
         public void Dispose()
         {
             foreach (var component in components)
+            {
+                component.Value.Release();
                 CoreUtils.Destroy(component.Value);
+            }
 
             components.Clear();
         }